Unfucked beat time.

+++ src/MidiGremlin/Instrument.cs
@@ -32,17 +32,17 @@ namespace MidiGremlin
             Play(_orchestra.CurrentTime(), music);
         }
 
-        public void Play (int startTime, MusicObject music)
+        public void Play (double startTime, MusicObject music)
         {
             _orchestra.CopyToOutput(new List<SingleBeat>(music.GetChildren(this, startTime)));
         }
 
-        public void Play(Tone tone, int duration, byte velocity = 64)
+        public void Play(Tone tone, double duration, byte velocity = 64)
         {
             Play(_orchestra.CurrentTime(), tone, duration,  velocity);
         }
 
-        public void Play(int startTime, Tone tone, int duration,  byte velocity = 64)
+        public void Play(double startTime, Tone tone, double duration,  byte velocity = 64)
         {
             Note note = new Note(tone, duration, velocity);
             Play(startTime, note);
